Rename stable coin for batch exchange
