+++ src/contracts/BatchExchangeContract.ts
@@ -58,7 +58,7 @@ export interface OrderPlacement {
 
 export interface BatchExchangeContract extends Contract {
   // constructor(jsonInterface: any[], address?: string, options?: ContractOptions)
-  // constructor (jsonInterface: AbiItem[], address?: string, options?: ContractOptions): StablecoinConverter;
+  // constructor (jsonInterface: AbiItem[], address?: string, options?: ContractOptions): BatchExchange;
 
   clone(): BatchExchangeContract
